requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
aioredis
sqlalchemy
asyncpg
alembic
psycopg2-binary
httpx
python-jose
pytest
pytest-asyncio
gunicorn
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation==0.40b0
requests
sqlalchemy
psycopg2-binary
